Switch to SwiftSyntaxMacrosGenericTestSupport
Replaces SwiftSyntaxMacrosTestSupport with
SwiftSyntaxMacrosGenericTestSupport and configures the failure handler
to use Testing.Issue.record to avoid XCTFail.
