+++ Tests/MMIOInterposableTests/MMIOTracingInterposer.swift
@@ -10,7 +10,6 @@
 //===----------------------------------------------------------------------===//
 
 import MMIOInterposable
-import MMIOUtilities
 import Testing
 
 class MMIOTracingInterposer {
@@ -63,36 +62,4 @@ extension MMIOTracingInterposer: MMIOInterposer {
 
     self.trace.append(.store(of: value, to: address))
   }
-}
-
-func assertMMIOAlignment<Value>(
-  pointer: UnsafePointer<Value>,
-  sourceLocation: SourceLocation = #_sourceLocation
-) where Value: FixedWidthInteger & UnsignedInteger {
-  let address = UInt(bitPattern: pointer)
-  let alignment = UInt(MemoryLayout<Value>.alignment)
-  #expect(
-    address.isMultiple(of: alignment),
-    """
-    Invalid load or store of type '\(Value.self)' from unaligned address \
-    '\(hex: address)'
-    """,
-    sourceLocation: sourceLocation)
-}
-
-func assertMMIOInterposerTrace(
-  interposer: MMIOTracingInterposer,
-  trace: [MMIOTracingInterposerEvent],
-  sourceLocation: SourceLocation = #_sourceLocation
-) {
-  // Exit early if the actual trace matches the expected trace.
-  let actualTrace = interposer.trace
-  let expectedTrace = trace
-  guard actualTrace != expectedTrace else { return }
-
-  Issue.record(
-    Comment(
-      rawValue: diff(
-        expected: expectedTrace, actual: actualTrace, noun: "trace")),
-    sourceLocation: sourceLocation)
 }
