+++ Tests/MMIOMacrosTests/Macros/RegisterBlockAndOffsetMacroTests.swift
@@ -12,7 +12,6 @@
 #if canImport(MMIOMacros)
 import SwiftSyntax
 import SwiftSyntaxMacros
-import SwiftSyntaxMacrosTestSupport
 import Testing
 
 @testable import MMIOMacros
